Fixed issue with flots and ints in the webapp

+++ kotlin-source/src/main/kotlin/com/termDeposits/api/ExampleApi.kt
@@ -115,7 +115,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
      */
     @POST
     @Path("issue_td")
-    fun issueTD(@QueryParam("td_value") tdValue: Int, @QueryParam("offering_institute") offeringInstitue:String,
+    fun issueTD(@QueryParam("td_value") tdValue: Float, @QueryParam("offering_institute") offeringInstitue:String,
                 @QueryParam("interest_percent") interestPercent: Float, @QueryParam("duration") duration: Int,
                 @QueryParam("customer_fname") firstName: String, @QueryParam("customer_lname") lastName: String,
                 @QueryParam("customer_anum") accountNum: String) : Response {
@@ -124,7 +124,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
         val kyc = KYC.KYCNameData(firstName, lastName, accountNum)
         //TODO use actual dates, for testing we use LocalDateTime.MIN for now
         val dateData = TermDeposit.DateData(LocalDateTime.MIN, duration)
-        val depositAmount = AMOUNT(tdValue, USD)
+        val depositAmount = AMOUNT(tdValue.toDouble(), USD)
 
         return try {
             val flowHandle = rpcOps.startTrackedFlow(IssueTD::Initiator, dateData, interestPercent, issuingInstitute, depositAmount, kyc)
@@ -142,7 +142,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
      */
     @POST
     @Path("activate_td")
-    fun activateTD(@QueryParam("td_value") tdValue: Int, @QueryParam("offering_institute") offeringInstitue:String,
+    fun activateTD(@QueryParam("td_value") tdValue: Float, @QueryParam("offering_institute") offeringInstitue:String,
                 @QueryParam("interest_percent") interestPercent: Float, @QueryParam("duration") duration: Int,
                 @QueryParam("customer_fname") firstName: String, @QueryParam("customer_lname") lastName: String,
                 @QueryParam("customer_anum") accountNum: String, @QueryParam("start_date") startDate: String,
@@ -153,7 +153,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
         val kyc = KYC.KYCNameData(firstName, lastName, accountNum)
         val startDateActual = LocalDateTime.parse(startDate+"T00:00:00")
         val dateData = TermDeposit.DateData(startDateActual, duration)
-        val depositAmount = AMOUNT(tdValue, USD)
+        val depositAmount = AMOUNT(tdValue.toDouble(), USD)
 
         return try {
             val flowHandle = rpcOps.startFlow(ActivateTD::Activator, dateData, interestPercent, issuingInstitute,
@@ -172,7 +172,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
      */
     @POST
     @Path("redeem_td")
-    fun redeemTD(@QueryParam("td_value") tdValue: Int, @QueryParam("offering_institute") offeringInstitue:String,
+    fun redeemTD(@QueryParam("td_value") tdValue: Float, @QueryParam("offering_institute") offeringInstitue:String,
                    @QueryParam("interest_percent") interestPercent: Float, @QueryParam("duration") duration: Int,
                    @QueryParam("customer_fname") firstName: String, @QueryParam("customer_lname") lastName: String,
                    @QueryParam("customer_anum") accountNum: String, @QueryParam("start_date") startDate: String) : Response {
@@ -181,7 +181,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
         val kyc = KYC.KYCNameData(firstName, lastName, accountNum)
         val startDateActual = LocalDateTime.parse(startDate+"T00:00:00")
         val dateData = TermDeposit.DateData(startDateActual, duration)
-        val depositAmount = AMOUNT(tdValue, USD)
+        val depositAmount = AMOUNT(tdValue.toDouble(), USD)
 
         return try {
             val flowHandle = rpcOps.startFlow(RedeemTD::RedemptionInitiator, dateData, interestPercent, issuingInstitute, depositAmount, kyc)
@@ -199,7 +199,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
      */
     @POST
     @Path("rollover_td")
-    fun rolloverTD(@QueryParam("td_value") tdValue: Int, @QueryParam("offering_institute") offeringInstitue:String,
+    fun rolloverTD(@QueryParam("td_value") tdValue: Float, @QueryParam("offering_institute") offeringInstitue:String,
                  @QueryParam("interest_percent") interestPercent: Float, @QueryParam("duration") duration: Int,
                  @QueryParam("customer_fname") firstName: String, @QueryParam("customer_lname") lastName: String,
                  @QueryParam("customer_anum") accountNum: String, @QueryParam("start_date") startDate: String,
@@ -212,7 +212,7 @@ class DepositsAPI(private val rpcOps: CordaRPCOps) {
         val kyc = KYC.KYCNameData(firstName, lastName, accountNum)
         val startDateActual = LocalDateTime.parse(startDate+"T00:00:00")
         val dateData = TermDeposit.DateData(startDateActual, duration)
-        val depositAmount = AMOUNT(tdValue, USD)
+        val depositAmount = AMOUNT(tdValue.toDouble(), USD)
 
         return try {
             val flowHandle = rpcOps.startFlow(RolloverTD::RolloverInitiator, dateData, interestPercent, issuingInstitute, depositAmount, rolloverTerms, kyc)
